+++ src/componants/loadingitems/techbadges.jsx
