+++ src/componants/shared/pointslist.jsx
@@ -5,7 +5,7 @@ import { Box, Stack, Typography } from '@mui/material'
 //*styles
 import "../../sass/shared/pointslist.scss"
 //*components
-import { WaitItemsSkelton } from '../loadinitems/pointslist'
+import { WaitItemsSkelton } from '../loadingitems/pointslist'
 
 export default function PointsList({ dir, loading, data=[] }) {
   return (
